Size adjustments for better fit to small probes
Hello and thank you for bringing landertrons back, they have been missed!

I propose those changes to the box, so that it becomes more useful for tiny probes. At the moment the box has the size of a probe core with reaction wheel and is quite heavy, EC hungry and pricy in comparison as well. So when building a small probe lander, it seems a bit off at the moment. With the new size it fits nicely on the side of standar probe core, like the HEX.